geneSurv

This function analyzes the ability of a gene to mark survival based on a robust version of the KM curves. The robust K-M estimator is obtained by a bootstrap strategy.

Usage

geneSurv(genExpr, time, status, geneName, boxplot = T, iter = 100, type = c("exprs", "risk"), cut_time = 10)

Arguments

genExpr

Vector with the normalized gene expression for each sample. names(genExpr) should contains the sample names.

time

Numeric vector containing the survival time for each sample in years, including the sample name in names(time).

status

Numeric vector with the status (censored 0 and not censored 1) for each sample. names(status) shoud include the sample names.

geneName

A character string with the name of the gene being analyzed.

boxplot

A logical value indicating whether to generate a boxplot of gene expression by survival group (default = TRUE).

iter

The number of iterations (bootstrap resampling) for calculating optimal group cutoffs (default = 100).

type

Defines if the KM curve groups are computed using risk ("risk") or gene expression (default "exprs").

cut_time

A numeric value specifying the cutoff time (in years) for survival analysis. All events beyond this time are treated as censored (default = 10 years).

Details

This function improves the stability and robustness of the K-M estimator using a bootstrap strategy. Patients are resampled with replacement giving rise to B replicates. The K-M estimator is obtained based on the replicates as well as the confidence intervals. The patients are stratified in two risk groups by an expression threshold that optimizes the log-rank statistics, that is the separability between the Kaplan-Meier curves for each group. This function implements a novel method to find the optimal threshold avoiding the problems of instability and unbalanced classes that suffer other implementations. Besides, a membership probability for each risk group is estimated from the classification of each sample in the replicates. This membership probability allow us to reclassify patients around the gene expression threshold in a more robust way. The function provides a robust estimation of the log-rank p-value and the Hazard ratio that allow us to evaluate the ability of a given gene to mark survival.

Value

Depending on the type run, the output changes.

For "type = exprs":

A Kaplan-Meier plot based on expression groups, a differential expression boxplot and a plot with the membership probability for each risk group. Additionally, an object with the following components:

geneName

A character string with the name of the gene being analyzed.

patientExpr

The expression level of each patient for the given gene.

patientClass

Vector of group classification according to the gene expression level: 2 = high expression and 1 = low expression level.

patientClassProbality

Vector of membership probabilities for the classification.

wilcox.pvalue

The p-value from the Wilcoxon test comparing the two expression groups.

plot_values

A list containing Kaplan-Meier fit results, log-rank p-value, and hazard ratio.

For "type = risk":

A Kaplan-Meier plot based on risk groups. Additionally, an object with the following components:

geneName

A character string with the name of the gene being analyzed.

patientExpr

The expression level of each patient for the given gene.

risk_score_predicted

A numeric vector of predicted relative risk scores for each patient.

plot_values

A list containing Kaplan-Meier fit results, log-rank p-value, and hazard ratio.

References

Martinez-Romero, J., Bueno-Fortes, S., Martín-Merino, M. et al. (2018) Survival marker genes of colorectal cancer derived from consistent transcriptomic profiling. BMC Genomics 19 (Suppl 8), 857 . doi:10.1186/s12864-018-5193-9

Santiago Bueno-Fortes, Alberto Berral-Gonzalez, José Manuel Sánchez-Santos, Manuel Martin-Merino, Javier De Las Rivas (2023) Identification of a gene expression signature associated with breast cancer survival and risk that improves clinical genomic platforms, Bioinformatics Advances, Volume 3, Issue 1, vbad037, doi:10.1093/bioadv/vbad037

Examples

  data(matrixData)
#> Warning: data set ‘matrixData’ not found
  
  set.seed(5)
  time <- mPheno$time
#> Error: object 'mPheno' not found
  names(time) <- rownames(mPheno)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'rownames': object 'mPheno' not found
  status <- mPheno$status
#> Error: object 'mPheno' not found
  names(status) <- rownames(mPheno)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'rownames': object 'mPheno' not found
  # The TIME value must be transformed to YEARS
  # The gene expression vector must be provided with the NAMES of each sample, that should match with the NAMES of the time and status vectors 
  outputKM <- geneSurv(mExprs[match("ESR1",rownames(mExprs)),], time, status, "ESR1", 
                       type = "exprs")
#> Error: object 'mExprs' not found
  
  # If we consider p53 gene instead (ENSG00000141510)
  outputKM.TP53 <- geneSurv(mExprs[match("TP53",rownames(mExprs)),], time, status, "TP53", 
                            type = "exprs")
#> Error: object 'mExprs' not found
                            
  # Generate the plots again
  ## Plots for c(type = exprs)
  asuri:::plotBoxplot(outputKM)
#> Error: object 'outputKM' not found
  asuri:::plotProbClass(outputKM)
#> Error: object 'outputKM' not found
  asuri:::plotKM(outputKM)
#> Error: object 'outputKM' not found
  ## Plots for c(type = risk)
  asuri:::plotKM(outputKM)
#> Error: object 'outputKM' not found