predSurvCurve

Function to predict the expected survival curve for a test patient given the risk score provided by a Cox regression model.

Usage

predSurvCurve( cox_pred_training, mSurv, cox_pred_test, eval_surv_times = NULL )

Arguments

cox_pred_training

A numeric vector representing the predicted risk scores (or log-risk scores) from the Cox model for the training set. It can be obtained from *predict.patientRisk* function.

mSurv

A data frame with two columns: "time" representing survival times, and "status" representing the event status (1 for event, 0 for censored) for the training dataset.

cox_pred_test

A numeric vector of length 1 representing the predicted risk score (or log-risk score) for the test patient. It can be obtained from *predict.patientRisk* function.

eval_surv_times

A numeric vector of times at which the survival curve should be evaluated. If NULL (default), the times will be taken from the training data up to the maximum event time.

Details

This function will obtain the expected survival curve for a test patient given the risk score estimated by a multivariate Cox regression model. The Breslow estimator is considered to approximate the baseline Hazard. The cumulative baseline survival function is smoothed using the Friedman's super smoother. The function interpolates linearly the survival probability for time points not in the training set.

Value

The function provides a plot for the expected survival curve. It shows also the median survival time, that is, the time at which the survival probability drops to 0.5. The function also returns a list with the following objects:

eval_times

Vector of times in years at which the survival curve has been evaluated.

S_0_t

Vector with the baseline survival probability estimated at each evaluation time.

S_test_patient

Vector with the survival probability estimated at each evaluation time for a given test patient.

References

Martinez-Romero, J., Bueno-Fortes, S., Martín-Merino, M. et al. (2018) Survival marker genes of colorectal cancer derived from consistent transcriptomic profiling. BMC Genomics 19 (Suppl 8), 857 . doi:10.1186/s12864-018-5193-9

Santiago Bueno-Fortes, Alberto Berral-Gonzalez, José Manuel Sánchez-Santos, Manuel Martin-Merino, Javier De Las Rivas (2023) Identification of a gene expression signature associated with breast cancer survival and risk that improves clinical genomic platforms, Bioinformatics Advances, Volume 3, Issue 1, vbad037, doi:10.1093/bioadv/vbad037

Examples

# Load the expression matrix with the subset of genes selected
data(matrixData)
#> Warning: data set ‘matrixData’ not found
set.seed(5)

# Train a multivariate risk predictor. If you have run already this function, please skip this step.
multivariate_risk_predictor <- patientRisk(mExprSelectedGenes, time, status)
#> Error: object 'mExprSelectedGenes' not found

# Expression matrix for a single test patient
i <- 5
mExprs_testData <- data.frame( mExprSelectedGenes[, i ] )
#> Error: object 'mExprSelectedGenes' not found
colnames( mExprs_testData ) <- colnames(mExprSelectedGenes)[ i ]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'colnames': object 'mExprSelectedGenes' not found

# Cox prediction for training and test sets
cox_pred_training <- as.vector( predict.patientRisk(multivariate_risk_predictor, mExprSelectedGenes )$risk_score )
#> Error: object 'mExprSelectedGenes' not found
cox_pred_test <- as.vector( predict.patientRisk(multivariate_risk_predictor, mExprs_testData )$risk_score )
#> Error: object 'mExprs_testData' not found

# Survival curve estimation
eval_surv_times <- seq(0, max(mPheno$time), by=0.2)
#> Error: object 'mPheno' not found
surv_curv_cox <- predSurvCurve( cox_pred_training, mPheno[,c(2,3)], cox_pred_test, eval_surv_times )
#> Error: object 'mPheno' not found