prefilterSAM

R function to pre-filter noisy genes, keeping the most relevant ones for microarrays or RNAseq data. These genes can be used later for further analysis.

Usage

prefilterSAM(mExpr, groups_vector, FDRfilter = 0.05, percentageFilter = 80, iter = 100)

Arguments

mExpr

A matrix where rows represent genes and columns represent samples. Each entry is the expression value of a gene in a sample.

groups_vector

A binary vector indicating the group assignment for each sample.

FDRfilter

A numeric value indicating the FDR threshold for selecting significant genes. Default is 0.05.

iter

The number of iterations for bootstrapping. Default is 100.

percentageFilter

A numeric value indicating the percentage of iterations a gene must appear in to be considered significant. Default is 80.

Details

This function implements SAM (Schwender H., 2022) robust diferential expression analysis based on bootstrap . It helps to remove noisy genes reducing the computational complexity of further analysis.

Value

An ordered vector with the names of differentially expressed genes between the categories of the grouping vector.

result

A list of DE genes ordered by SAM d.value and filtered by percentageFilter.

References

Schwender H (2022). _siggenes: Multiple Testing using SAM and Efron's Empirical Bayes Approaches_. R package version 1.72.0.

Martinez-Romero, J., Bueno-Fortes, S., Martín-Merino, M. et al. (2018) Survival marker genes of colorectal cancer derived from consistent transcriptomic profiling. BMC Genomics 19 (Suppl 8), 857 . doi:10.1186/s12864-018-5193-9

Santiago Bueno-Fortes, Alberto Berral-Gonzalez, José Manuel Sánchez-Santos, Manuel Martin-Merino, Javier De Las Rivas (2023) Identification of a gene expression signature associated with breast cancer survival and risk that improves clinical genomic platforms, Bioinformatics Advances, Volume 3, Issue 1, vbad037, doi:10.1093/bioadv/vbad037

Examples

  # Bootstrapped differential expression based on SAM.
  # Parameters: FDR = 0.05, iter = 100, percentage filter = 80 %
  # CAUTION: if the data have a high number of genes this function will take several minutes to compute.
  data(prefilterSAM)

  set.seed(5)
  DE_list_genes <- prefilterSAM(mExprs, mPheno$ER.IHC)
#> 2025-04-24 14:05:19.184687
#> Progress: 1/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.882082, all FDRs are smaller than 0.05.
#> Progress: 2/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.767022, all FDRs are smaller than 0.05.
#> Progress: 3/100
#> For Delta >= 0.299952, all FDRs are smaller than 0.05.
#> Progress: 4/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.710353, all FDRs are smaller than 0.05.
#> Progress: 5/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.747871, all FDRs are smaller than 0.05.
#> Progress: 6/100
#> For Delta >= 0.152733, all FDRs are smaller than 0.05.
#> Progress: 7/100
#> For Delta >= 0.1, all FDRs are smaller than 0.05.
#> Progress: 8/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.114115, all FDRs are smaller than 0.05.
#> Progress: 9/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.159556, all FDRs are smaller than 0.05.
#> Progress: 10/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.917052, all FDRs are smaller than 0.05.
#> Progress: 11/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.992159, all FDRs are smaller than 0.05.
#> Progress: 12/100
#> For Delta >= 0.1, all FDRs are smaller than 0.05.
#> Progress: 13/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.57915, all FDRs are smaller than 0.05.
#> Progress: 14/100
#> For Delta >= 0.271571, all FDRs are smaller than 0.05.
#> Progress: 15/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.186251, all FDRs are smaller than 0.05.
#> Progress: 16/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.090498, all FDRs are smaller than 0.05.
#> Progress: 17/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.780918, all FDRs are smaller than 0.05.
#> Progress: 18/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.565798, all FDRs are smaller than 0.05.
#> Progress: 19/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.499581, all FDRs are smaller than 0.05.
#> Progress: 20/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.926352, all FDRs are smaller than 0.05.
#> Progress: 21/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.372193, all FDRs are smaller than 0.05.
#> Progress: 22/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.271911, all FDRs are smaller than 0.05.
#> Progress: 23/100
#> For Delta >= 0.288377, all FDRs are smaller than 0.05.
#> Progress: 24/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.1, all FDRs are smaller than 0.05.
#> Progress: 25/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.441612, all FDRs are smaller than 0.05.
#> Progress: 26/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.452891, all FDRs are smaller than 0.05.
#> Progress: 27/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.490172, all FDRs are smaller than 0.05.
#> Progress: 28/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.751042, all FDRs are smaller than 0.05.
#> Progress: 29/100
#> For Delta >= 0.193117, all FDRs are smaller than 0.05.
#> Progress: 30/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.433616, all FDRs are smaller than 0.05.
#> Progress: 31/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.303926, all FDRs are smaller than 0.05.
#> Progress: 32/100
#> For Delta >= 0.1, all FDRs are smaller than 0.05.
#> Progress: 33/100
#> For Delta >= 0.136389, all FDRs are smaller than 0.05.
#> Progress: 34/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 2.220244, all FDRs are smaller than 0.05.
#> Progress: 35/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.634363, all FDRs are smaller than 0.05.
#> Progress: 36/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.430684, all FDRs are smaller than 0.05.
#> Progress: 37/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.197038, all FDRs are smaller than 0.05.
#> Progress: 38/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.84943, all FDRs are smaller than 0.05.
#> Progress: 39/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.456953, all FDRs are smaller than 0.05.
#> Progress: 40/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.620601, all FDRs are smaller than 0.05.
#> Progress: 41/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.745684, all FDRs are smaller than 0.05.
#> Progress: 42/100
#> For Delta >= 0.161207, all FDRs are smaller than 0.05.
#> Progress: 43/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.97555, all FDRs are smaller than 0.05.
#> Progress: 44/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.395494, all FDRs are smaller than 0.05.
#> Progress: 45/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.1, all FDRs are smaller than 0.05.
#> Progress: 46/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.865544, all FDRs are smaller than 0.05.
#> Progress: 47/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.695939, all FDRs are smaller than 0.05.
#> Progress: 48/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.329712, all FDRs are smaller than 0.05.
#> Progress: 49/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.155021, all FDRs are smaller than 0.05.
#> Progress: 50/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.1, all FDRs are smaller than 0.05.
#> Progress: 51/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.759343, all FDRs are smaller than 0.05.
#> Progress: 52/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.25334, all FDRs are smaller than 0.05.
#> Progress: 53/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.604431, all FDRs are smaller than 0.05.
#> Progress: 54/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.864314, all FDRs are smaller than 0.05.
#> Progress: 55/100
#> For Delta >= 0.1, all FDRs are smaller than 0.05.
#> Progress: 56/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.399625, all FDRs are smaller than 0.05.
#> Progress: 57/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.400041, all FDRs are smaller than 0.05.
#> Progress: 58/100
#> For Delta >= 0.293107, all FDRs are smaller than 0.05.
#> Progress: 59/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.854155, all FDRs are smaller than 0.05.
#> Progress: 60/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.565921, all FDRs are smaller than 0.05.
#> Progress: 61/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.738672, all FDRs are smaller than 0.05.
#> Progress: 62/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.607098, all FDRs are smaller than 0.05.
#> Progress: 63/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.600248, all FDRs are smaller than 0.05.
#> Progress: 64/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.548246, all FDRs are smaller than 0.05.
#> Progress: 65/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.028952, all FDRs are smaller than 0.05.
#> Progress: 66/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.145446, all FDRs are smaller than 0.05.
#> Progress: 67/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.489462, all FDRs are smaller than 0.05.
#> Progress: 68/100
#> For Delta >= 0.1, all FDRs are smaller than 0.05.
#> Progress: 69/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.1332, all FDRs are smaller than 0.05.
#> Progress: 70/100
#> For Delta >= 0.304696, all FDRs are smaller than 0.05.
#> Progress: 71/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.511565, all FDRs are smaller than 0.05.
#> Progress: 72/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.862977, all FDRs are smaller than 0.05.
#> Progress: 73/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.698075, all FDRs are smaller than 0.05.
#> Progress: 74/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.308615, all FDRs are smaller than 0.05.
#> Progress: 75/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.423496, all FDRs are smaller than 0.05.
#> Progress: 76/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.050901, all FDRs are smaller than 0.05.
#> Progress: 77/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.22891, all FDRs are smaller than 0.05.
#> Progress: 78/100
#> For Delta >= 0.387276, all FDRs are smaller than 0.05.
#> Progress: 79/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.110442, all FDRs are smaller than 0.05.
#> Progress: 80/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.339186, all FDRs are smaller than 0.05.
#> Progress: 81/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.662679, all FDRs are smaller than 0.05.
#> Progress: 82/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.000853, all FDRs are smaller than 0.05.
#> Progress: 83/100
#> For Delta >= 0.105715, all FDRs are smaller than 0.05.
#> Progress: 84/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.019102, all FDRs are smaller than 0.05.
#> Progress: 85/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.298774, all FDRs are smaller than 0.05.
#> Progress: 86/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.234585, all FDRs are smaller than 0.05.
#> Progress: 87/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.951315, all FDRs are smaller than 0.05.
#> Progress: 88/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.42788, all FDRs are smaller than 0.05.
#> Progress: 89/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.234133, all FDRs are smaller than 0.05.
#> Progress: 90/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.539187, all FDRs are smaller than 0.05.
#> Progress: 91/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.28385, all FDRs are smaller than 0.05.
#> Progress: 92/100
#> For Delta >= 0.150867, all FDRs are smaller than 0.05.
#> Progress: 93/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.985415, all FDRs are smaller than 0.05.
#> Progress: 94/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.774216, all FDRs are smaller than 0.05.
#> Progress: 95/100
#> For Delta >= 0.466054, all FDRs are smaller than 0.05.
#> Progress: 96/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 0.324306, all FDRs are smaller than 0.05.
#> Progress: 97/100
#> For Delta >= 0.1, all FDRs are smaller than 0.05.
#> Progress: 98/100
#> For Delta >= 0.421601, all FDRs are smaller than 0.05.
#> Progress: 99/100
#> For Delta >= 0.203498, all FDRs are smaller than 0.05.
#> Progress: 100/100
#> Warning: The spline based estimation of pi0 results in a non-positive value of pi0.
#> Therefore, pi0 is estimated by using lambda = 0.5.
#> For Delta >= 1.217753, all FDRs are smaller than 0.05.